Minor enhancements.

+++ hacking_arduino/part2/ParkDistanceControl/infrared_sensor.cpp
@@ -5,7 +5,7 @@
 
 namespace arduino {
   namespace sensors {
-    InfraredSensor::InfraredSensor(const uint16_t pin) 
+    InfraredSensor::InfraredSensor(const uint8_t pin) 
       : _pin(pin) 
     {
       for (uint16_t i = 0; i < _buffer.getBufferSize(); i++)
